sh/pre_config/requirements.txt
Babel==2.7.0
pbr==5.4.1
ply==3.11
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
yaql==1.1.3
